refacting away from helper paradigm
